refactor : 코드 구조 조정
