+++ src/main/resources/templates/post/list.html
@@ -6,7 +6,8 @@
 </head>
 <body>
 
+<div>hello~ this page shows list of posts </div>
 
-<script src="/static/js/list.js"></script>
+<script src="/js/list.js"></script>
 </body>
 </html>
